[dzil] add github_user overrides; use GitHubMeta plugin
Anything using this bundle is liable to be a package I own -- including
packages others may upload to the CPAN.  In the case that it's not, we setting
set_github_user to false will defer to the plugin's judgment, while setting
github_user to a different userid will cause that id to be used instead of
mine (e.g. for an organization or the like).